+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Reworked UI
 * Store configs in ~/Library/Preferences on Mac
 * Make annotationDBMap public for mods to use
+* Add some functionality to SpireConfig
 
 #### v2.9.1 ####
 * Patch to always enable Custom mode